separated monster and player concerns into their separate files, finished add monster functionality, and added to summary page
