+++ assets/js/createEncounter.js
@@ -1,18 +1,31 @@
-import { Pc } from './pc.js';
+import { printPc, createNewPc } from './selectPc.js';
+import { printMonsters, searchMonster } from './selectMonster.js';
 
 const headerTitle = document.querySelector('#header-title');
 const mainEl = document.querySelector('main');
 const footerNav = document.querySelector('#footer-navigation');
 
 let pageIndex = 0;
-let stagedArr = [];
+// let stagedMon = [];
 let storedPc = [];
+let storedMonster = [];
 
-const checkStoredPc = () => {
+export const checkStoredPc = () => {
 	storedPc = JSON.parse(localStorage.getItem('playerCharacter'));
 	if (storedPc == null) {
 		storedPc = [];
 	}
+
+	return storedPc;
+};
+
+export const checkStoredMonster = () => {
+	storedMonster = JSON.parse(localStorage.getItem('savedMonster'));
+	if (storedMonster == null) {
+		storedMonster = [];
+	}
+
+	return storedMonster;
 };
 
 // renders the create encounter section of the application
@@ -42,6 +55,7 @@ export const createEncounter = () => {
 	mainEl.appendChild(selectionDiv);
 
 	checkStoredPc();
+	checkStoredMonster();
 
 	// uses current pageIndex to dynamically change what is displayed
 	switch (pageIndex) {
@@ -50,46 +64,21 @@ export const createEncounter = () => {
 			addBtn.textContent = '+ add new PC';
 			addBtn.addEventListener('click', createNewPc);
 
-			for (let i = 0; i < storedPc.length; i++) {
-				createCharacterRow(
-					storedPc[i],
-					i // this is used for dataset index
-				);
-			}
+			printPc();
 
 			break;
 
 		case 1: // * monsters
-			mainEl.textContent = '';
+			text.textContent = 'Select your Monsters:';
+			addBtn.textContent = '+ search for a monster';
+			addBtn.addEventListener('click', searchMonster);
+
 			printMonsters();
+
 			break;
 
 		default:
 			console.log('error');
-	}
-
-	// handles close button on each row
-	let closeBtns = document.querySelectorAll('.close');
-
-	for (let i = 0; i < closeBtns.length; i++) {
-		closeBtns[i].addEventListener('click', function (e) {
-			e.stopPropagation();
-
-			if (confirm('Are you sure you want to delete?')) {
-				// remove from page
-				this.parentElement.parentElement.remove();
-				// remove from local storage
-				let index = this.parentElement.parentElement.dataset.index;
-				let pcArray = JSON.parse(
-					localStorage.getItem('playerCharacter')
-				);
-				pcArray.splice(index, 1);
-				localStorage.setItem(
-					'playerCharacter',
-					JSON.stringify(pcArray)
-				);
-			}
-		});
 	}
 
 	// navigation buttons
@@ -132,46 +121,12 @@ export const createEncounter = () => {
 	footerNav.append(prevBtn, nextBtn);
 };
 
-const createCharacterRow = (object, index) => {
-	let selectionDiv = document.querySelector('#selection-div');
-
-	let row = `
-        <div class="player-level col border bg-light d-flex flex-column justify-content-center align-items-center">
-            <p class="m-0 text-dark">Lvl</p>
-            <p class="m-0 text-dark">${object.pcLevel}</p>
-        </div>
-        <div class="row-title col-9 border d-flex align-items-center">
-            <div>
-            <h5 class="m-0">${object.pcName}</h5>
-            <p class="m-0">${object.pcClass}</p>
-            </div>
-            <span class="close">x</span>
-        </div>
-	`;
-
-	let div = document.createElement('div');
-	div.classList.add('row', 'justify-content-center', 'my-2');
-	div.dataset.index = index;
-	div.insertAdjacentHTML('beforeend', row);
-
-	// stages a character for the encounter
-	div.addEventListener('click', function () {
-		if (div.classList.contains('selected')) {
-			div.classList.remove('selected');
-			stagedArr.splice(
-				stagedArr.findIndex((e) => e.pcName === object.pcName),
-				1
-			);
-		} else {
-			div.classList.add('selected');
-			stagedArr.push(object);
-		}
-	});
-
-	selectionDiv.append(div);
-};
+// ! ####### SUMMARY #######
 
 const printSummary = () => {
+	storedPc = checkStoredPc();
+	storedMonster = checkStoredMonster();
+
 	headerTitle.textContent = 'ENCOUNTER SUMMARY';
 
 	// main content
@@ -182,17 +137,32 @@ const printSummary = () => {
 	pcHeader.textContent = "PC's:";
 	pcDiv.append(pcHeader);
 
-	let npcDiv = document.createElement('div');
-	let npcHeader = document.createElement('h4');
-	npcHeader.textContent = "NPC's:";
-	npcDiv.append(npcHeader);
+	let pcUl = document.createElement('ul');
+	storedPc.forEach((pc) => {
+		let li = document.createElement('li');
+		li.textContent = `${pc.pcName} (${pc.pcClass}) -- Lvl. ${pc.pcLevel}`;
+
+		pcUl.append(li);
+	});
+	pcDiv.append(pcUl);
+
+	// -----------------------------------------
 
 	let monDiv = document.createElement('div');
 	let monHeader = document.createElement('h4');
 	monHeader.textContent = "monsters's:";
 	monDiv.append(monHeader);
 
-	summaryContent.append(pcDiv, npcDiv, monDiv);
+	let monUl = document.createElement('ul');
+	storedMonster.forEach((mon) => {
+		let li = document.createElement('li');
+		li.textContent = `${mon.name}`;
+
+		monUl.append(li);
+	});
+	monDiv.append(monUl);
+
+	summaryContent.append(pcDiv, monDiv);
 	mainEl.append(summaryContent);
 
 	// bottom buttons
@@ -244,175 +214,4 @@ const printSummary = () => {
 	});
 
 	footerNav.append(saveBtn, runBtn, difficultyMeter, prevBtn, clearBtn);
-};
-
-const createNewPc = () => {
-	mainEl.textContent = '';
-	footerNav.textContent = '';
-	headerTitle.textContent = 'CREATE NEW PC';
-
-	let form = `
-    <form autocomplete="off">
-        <div class="mb-3">
-            <label for="pcNameInput" class="form-label">PC Name:</label>
-            <input type=text class="form-control" id="pcNameInput" placeholder="Type a name...">
-        </div>
-        <div class="mb-3">
-            <label for="pcClassInput" class="form-label">PC Class:</label>
-            <select class="form-select" id="pcClassInput" aria-label="Possible classes">
-                <option selected>Choose a class</option>
-                <option value="🪓 Barbarian"> 🪓 Barbarian</option>
-                <option value="🎼 Bard"> 🎼 Bard</option>
-                <option value="🙏 Cleric">🙏 Cleric</option>
-                <option value="🌱 Druid">🌱 Druid</option>
-                <option value="🤺 Fighter">🤺 Fighter</option>
-                <option value="👊 Monk">👊 Monk</option>
-                <option value="⛪ Paladin">⛪ Paladin</option>
-                <option value="🏹 Ranger">🏹 Ranger</option>
-                <option value="🔪 Rogue">🔪 Rogue</option>
-                <option value="🎇 Sorcerer">🎇 Sorcerer</option>
-                <option value="👿 Warlock">👿 Warlock</option>
-                <option value="📔 Wizard">📔 Wizard</option>
-                <option value="🔨 Artificer">🔨 Artificer</option>
-                <option value="❓ Other">❓ Other</option>
-            </select>
-        </div>
-        <div class="mb-3">
-            <label for"pcLevelInput" class"form-label">PC Level:</label>
-            <select class="form-select" id="pcLevelInput" aria-label="PC Level">
-                <option selected>Choose a level</option>
-                <option value="1">1</option>
-                <option value="2">2</option>
-                <option value="3">3</option>
-                <option value="4">4</option>
-                <option value="5">5</option>
-                <option value="6">6</option>
-                <option value="7">7</option>
-                <option value="8">8</option>
-                <option value="9">9</option>
-                <option value="10">10</option>
-                <option value="11">11</option>
-                <option value="12">12</option>
-                <option value="13">13</option>
-                <option value="14">14</option>
-                <option value="15">15</option>
-                <option value="16">16</option>
-                <option value="17">17</option>
-                <option value="18">18</option>
-                <option value="19">19</option>
-                <option value="20">20</option>
-            </select>
-        </div>
-    </form>
-    `;
-
-	mainEl.innerHTML = form;
-
-	let backBtn = document.createElement('button');
-	backBtn.textContent = 'Back';
-	backBtn.classList.add('btn', 'btn-secondary', 'btn-lg', 'mb-2', 'col-12');
-	backBtn.addEventListener('click', function () {
-		createEncounter();
-	});
-
-	let addPcBtn = document.createElement('button');
-	addPcBtn.textContent = 'Add PC';
-	addPcBtn.classList.add('btn', 'btn-success', 'btn-lg', 'mb-1', 'col-12');
-	addPcBtn.addEventListener('click', function () {
-		let pcName = document.querySelector('#pcNameInput').value;
-		let pcClass = document.querySelector('#pcClassInput').value;
-		let pcLevel = document.querySelector('#pcLevelInput').value;
-
-		const player = new Pc(pcName, pcClass, pcLevel);
-		storedPc.push(player);
-		localStorage.setItem('playerCharacter', JSON.stringify(storedPc));
-
-		createEncounter();
-	});
-
-	footerNav.append(addPcBtn, backBtn);
-};
-
-const printMonsters = () => {
-	// container to contain monster search
-	let div = document.createElement('div');
-
-	let searchInputEl = `
-    <form>
-    <label for="monster-search" class="form-label">Search</label>
-    <input type="text" class="form-control mb-3" id="monster-search">
-    <a id="monster-search-btn" class="btn btn-primary btn-block">Search...</a>
-    </form>
-    `;
-
-	div.insertAdjacentHTML('beforeend', searchInputEl);
-
-	mainEl.append(div);
-
-	// container to hold monster results
-	let monsterDiv = document.createElement('div');
-	monsterDiv.classList.add('container', 'mt-4');
-	monsterDiv.setAttribute('id', 'monster-div');
-
-	mainEl.appendChild(monsterDiv);
-
-	let searchBtn = document.querySelector('#monster-search-btn');
-	let searchInput = document.querySelector('#monster-search');
-
-	searchBtn.addEventListener('click', () => {
-		fetchMonsters(searchInput.value);
-	});
-};
-
-// calls the Open5e API
-const fetchMonsters = (search) => {
-	fetch(`https://api.open5e.com/monsters/?search=${search}`)
-		.then((response) => response.json())
-		.then((data) => {
-			// console.log(data.results);
-			for (let i = 0; i < data.results.length; i++) {
-				createMonsterRow(data.results[i]);
-			}
-		});
-};
-
-const createMonsterRow = (monster) => {
-	console.log(monster);
-	let monsterDiv = document.querySelector('#monster-div');
-
-	let row = `
-	    <div class="enemy-number col border bg-light d-flex flex-column justify-content-center align-items-center">
-            <form>
-                <label for="monster-num" class="form-label">#</label>
-                <input type="number" id="monster-num" class="form-control">
-            </form>
-	    </div>
-	    <div class="row-title col-9 border d-flex align-items-center">
-	        <div>
-	        <h5 class="m-0">${monster.name}</h5>
-	        <p class="m-0">${monster.subtype}</p>
-	        </div>
-	        <span class="close">x</span>
-	    </div>
-	`;
-
-	let div = document.createElement('div');
-	div.classList.add('row', 'justify-content-center', 'my-2');
-	div.insertAdjacentHTML('beforeend', row);
-
-	// // stages a character for the encounter
-	// div.addEventListener('click', function () {
-	// 	if (div.classList.contains('selected')) {
-	// 		div.classList.remove('selected');
-	// 		stagedArr.splice(
-	// 			stagedArr.findIndex((e) => e.pcName === object.pcName),
-	// 			1
-	// 		);
-	// 	} else {
-	// 		div.classList.add('selected');
-	// 		stagedArr.push(object);
-	// 	}
-	// });
-
-	monsterDiv.append(div);
 };
