+++ assets/js/monster.js
@@ -1,0 +1,7 @@
+export class Monster {
+	constructor(name, hp, ac) {
+		this.name = name;
+		this.hp = hp;
+		this.ac = ac;
+	}
+}
